Adds exceptions to Service methods in BookingService

+++ backend/src/main/java/io/bookitnow/backend/v1/service/BookingService.java
@@ -4,9 +4,12 @@ import io.bookitnow.backend.v1.DTOs.requests.BookingRequest;
 import io.bookitnow.backend.v1.DTOs.responses.BookingResponse;
 import io.bookitnow.backend.v1.entity.Booking;
 import io.bookitnow.backend.v1.repository.BookingRepository;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @Service
 public class BookingService {
@@ -45,32 +48,52 @@ public class BookingService {
         return bookings.stream().map(this::mapToResponse).toList();
     }
 
-    public BookingResponse getBookingById(Long id) {
-        Booking booking = bookingRepository.findById(id).orElseThrow();
+    public BookingResponse getBookingById(@NotNull Long id) {
+        if (id < 1) {
+            throw new IllegalArgumentException("Invalid booking id");
+        }
+        Booking booking = bookingRepository.findById(id).orElseThrow(() -> new NoSuchElementException("Booking not found"));
         return mapToResponse(booking);
     }
 
-    public BookingResponse createBooking(BookingRequest bookingRequest) {
-        Booking booking = mapToBooking(bookingRequest);
-        bookingRepository.save(booking);
-        return mapToResponse(booking);
+    public BookingResponse createBooking(@Valid BookingRequest bookingRequest) {
+        try {
+            Booking booking = mapToBooking(bookingRequest);
+            bookingRepository.save(booking);
+            return mapToResponse(booking);
+        } catch (Exception e) {
+            throw new IllegalArgumentException("Invalid booking request");
+        }
     }
 
-    public BookingResponse updateBooking(Long id, BookingRequest bookingRequest) {
-        Booking booking = bookingRepository.findById(id).orElseThrow();
-        booking.setUserId(bookingRequest.getUserId());
-        booking.setBookingDateTImeStart(bookingRequest.getBookingDateTimeStart());
-        booking.setBookingDateTimeEnd(bookingRequest.getBookingDateTimeEnd());
-        booking.setDurationInMinutes(bookingRequest.getDurationInMinutes());
-        booking.setIsCancelled(bookingRequest.getIsCancelled());
-        bookingRepository.save(booking);
-        return mapToResponse(booking);
+    public BookingResponse updateBooking(@NotNull Long id, @Valid BookingRequest bookingRequest) {
+        if (id < 1) {
+            throw new IllegalArgumentException("Invalid booking id");
+        }
+        Booking booking = bookingRepository.findById(id).orElseThrow(() -> new NoSuchElementException("Booking not found"));
+        try {
+            booking.setUserId(bookingRequest.getUserId());
+            booking.setBookingDateTImeStart(bookingRequest.getBookingDateTimeStart());
+            booking.setBookingDateTimeEnd(bookingRequest.getBookingDateTimeEnd());
+            booking.setDurationInMinutes(bookingRequest.getDurationInMinutes());
+            booking.setIsCancelled(bookingRequest.getIsCancelled());
+            bookingRepository.save(booking);
+            return mapToResponse(booking);
+        } catch (Exception e) {
+            throw new IllegalArgumentException("Invalid booking request");
+        }
     }
 
     public void deleteBooking(Long id) {
-        bookingRepository.deleteById(id);
+        if (id < 1) {
+            throw new IllegalArgumentException("Invalid booking id");
+        }
+        try {
+            bookingRepository.deleteById(id);
+        } catch (Exception e) {
+            throw new NoSuchElementException("Booking not found");
+        }
     }
-
 
 
 }
